app/consumer/requirements.txt
pika==1.3.2
redis==5.0.1
psycopg2-binary==2.9.9 